requirements.txt
fastapi
uvicorn[standard]
pandas
shapely
requests
python-dotenv